refactor: changed user model & added user functionality
