+++ backend/controllers/login.controller.js
@@ -1,28 +1,6 @@
 const { User } = require("../models/user.model");
 const bcrypt = require("bcrypt");
-
-const signUpNewUser = async (req, res, next) => {
-  try {
-    const { email, password, name } = req.body;
-    let user = await User.findOne({ email });
-    if (user) {
-      return res
-        .status(403)
-        .json({ success: false, errorMessage: "User already exists!" });
-    }
-
-    user = new User({ name, email, password });
-    const salt = await bcrypt.genSalt(10);
-    const newPassword = await bcrypt.hash(password, salt);
-    user.set("password", newPassword);
-    await user.save();
-    res
-      .status(201)
-      .json({ success: true, message: "User registration successful" });
-  } catch (error) {
-    next(error);
-  }
-};
+const jwt = require("jsonwebtoken");
 
 const verifyUserCredentials = async (req, res, next) => {
   try {
@@ -43,7 +21,10 @@ const verifyUserCredentials = async (req, res, next) => {
         },
         process.env["SECRET"]
       );
-      return res.status(200).json({ success: true, token, name: user.name });
+      const { firstname, lastname } = user;
+      return res
+        .status(201)
+        .json({ success: true, token, firstname, lastname });
     }
 
     res
@@ -54,4 +35,4 @@ const verifyUserCredentials = async (req, res, next) => {
   }
 };
 
-module.exports = { signUpNewUser, verifyUserCredentials };
+module.exports = { verifyUserCredentials };
