+++ backend/routes/createUser.router.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router({ mergeParams: true });
+const { signUpNewUser } = require("../controllers/user.controller");
+
+router.route("/").post(signUpNewUser);
+
+module.exports = router;
