+++ backend/routes/login.router.js
@@ -1,12 +1,7 @@
 const express = require("express");
 const router = express.Router({ mergeParams: true });
-const { User } = require("../models/user.model");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-const { signUpNewUser, verifyUserCredentials } = require("../controllers/login.controller");
+const { verifyUserCredentials } = require("../controllers/login.controller");
 
-router.route("/").post(signUpNewUser);
-
-router.route("/login").post(verifyUserCredentials);
+router.route("/").post(verifyUserCredentials);
 
 module.exports = router;
