+++ backend/routes/user.router.js
@@ -1,7 +1,5 @@
 const express = require("express");
 const router = express.Router({ mergeParams: true });
-const { User } = require("../models/user.model");
-const bcrypt = require("bcrypt");
 const {
   getUserDetails,
   changePassword,
